Rename `SectionManager` -> `SectionDirector`
