+++ pack/rel_objects.txt
@@ -51,7 +51,7 @@ out/Page.o
 out/rel/text_80602d2c_80621d0c.o
 out/Section.o
 out/rel/text_80624690_80634b80.o
-out/SectionManager.o
+out/SectionDirector.o
 out/rel/text_80635eb8_80786e60.o
 out/BoxColManager.o
 out/rel/rodata_80896158_808a4800.o
