+++ pack/rel_slices.csv
@@ -16,6 +16,6 @@ enabled,name,textStart,textEnd,ctorsStart,ctorsEnd,dtorsStart,dtorsEnd,rodataSta
 1,source/game/ui/MessageGroup.cpp,0x805f8b34,0x805f8d5c,,,,,0x80896148,0x80896158,,,,
 1,source/game/ui/Page.cpp,0x8060197c,0x80602d2c,,,,,,,,,,
 1,source/game/ui/Section.cpp,0x80621D0C,0x80624690,,,,,,,,,,
-1,source/game/ui/SectionManager.cpp,0x80634B80,0x80635EB8,,,,,,,,,,
+1,source/game/ui/SectionDirector.cpp,0x80634B80,0x80635EB8,,,,,,,,,,
 1,source/game/geo/BoxColManager.cpp,0x80786E60,0x80786ED0,,,,,,,,,,
 1,source/game/geo/BoxColUnit.cpp,0x80786ED0,0x80786F98,,,,,0x808A4800,0x808A4804,,,,
