+++ sources.py
@@ -458,7 +458,7 @@ SOURCES_REL = [
     Source(src="source/game/ui/UIControl.cpp", cc='4201_127', opts=REL_OPTS),
     Source(src="source/game/ui/Page.cpp", cc='4201_127', opts=REL_OPTS),
     Source(src="source/game/ui/Section.cpp", cc='4201_127', opts=REL_OPTS),
-    Source(src="source/game/ui/SectionManager.cpp", cc='4201_127', opts=REL_OPTS),
+    Source(src="source/game/ui/SectionDirector.cpp", cc='4201_127', opts=REL_OPTS),
     Source(src="source/game/system/CourseMap.cpp", cc='4201_127', opts=REL_OPTS),
     Source(src="source/game/system/DvdArchive.cpp", cc='4201_127', opts=REL_OPTS),
     Source(src="source/game/system/GhostFile.cpp", cc='4201_127', opts=REL_OPTS),
